requirements.txt
streamlit
pdfminer.six
python-docx
openai